Added panning and zooming
If you update the preprocessor define to build the game with more rows and columns, it is much easier to play if you can zoom and pan.

If your camera gets in a weird spot, you can press 0 to reset it.

This change in scale however really broke the text drawing to show the neighboring bombs, so I also added a sprite atlas of pre-drawn tiles that are drawn for revealed tiles which have a value for neighboring bombs.